refactor: consolidate skaff-lib test helpers
